+++ src/index.ts
@@ -1,12 +1,12 @@
 import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
 
 import instanceConfigFields from './instanceConfigFields';
-import fetchFindings from './steps/fetchFindings';
+import { steps } from './steps/fetchFindings';
 import { IntegrationConfig } from './types';
 import validateInvocation from './validateInvocation';
 
 export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = {
   instanceConfigFields,
   validateInvocation,
-  integrationSteps: [fetchFindings],
+  integrationSteps: steps,
 };