+++ src/converters.ts
@@ -1,10 +1,12 @@
 import {
   Entity,
+  parseTimePropertyValue,
   Relationship,
   RelationshipDirection,
 } from '@jupiterone/integration-sdk-core';
+import { Entities, Relationships } from './constants';
 
-import { FindingEntity } from './types';
+import { AggregatedIssue, CVEEntity, CWEEntity, FindingEntity } from './types';
 
 export const SNYK_SERVICE_ENTITY_TYPE = 'snyk_account';
 export const SNYK_FINDING_ENTITY_TYPE = 'snyk_finding';
@@ -18,110 +20,57 @@ export const SNYK_FINDING_CWE_RELATIONSHIP_TYPE = 'snyk_finding_exploits_cwe';
 
 const CVE_URL_BASE = 'https://nvd.nist.gov/vuln/detail/';
 
-function getTime(time: Date | string): number {
-  return new Date(time).getTime();
-}
-
-export interface SnykVulnIssue {
-  id: string;
-  url: string;
-  title: string;
-  type: string;
-  description: string;
-  from: string[];
-  package: string;
-  version: string;
-  severity: string;
-  language: string;
-  packageManager: string;
-  publicationTime: Date;
-  disclosureTime: Date;
-  isUpgradable: string;
-  isPatchable: string;
-  identifiers: Identifier;
-  cvssScore: number;
-  patches: Patch[];
-  upgradePath: string[];
-}
-
-export interface Identifier {
-  CVE: string[];
-  CWE: string[];
-}
-
-export interface Patch {
-  id: string;
-  urls: string[];
-  version: string;
-  comments: string[];
-  modificationTime: Date;
-}
-
-export interface Project {
-  name: string;
-  id: string;
-  createdOn: Date;
-  origin: string;
-  totalDependencies: number;
-  issueCountsBySeverity: IssueCount;
-}
-
-export interface IssueCount {
-  low: number;
-  medium: number;
-  high: number;
-}
-
 export function createServiceEntity(orgId: string): Entity {
   return {
     _key: `snyk:${orgId}`,
-    _type: SNYK_SERVICE_ENTITY_TYPE,
-    _class: ['Service', 'Account'],
+    _type: Entities.SNYK_ACCOUNT._type,
+    _class: Entities.SNYK_ACCOUNT._class,
     category: 'code dependency scan',
     displayName: `snyk/${orgId}`,
   };
 }
 
-export function createFindingEntity(vuln: SnykVulnIssue): FindingEntity {
+export function createFindingEntity(vuln: AggregatedIssue): FindingEntity {
   return {
-    _class: 'Finding',
+    _class: Entities.SNYK_FINDING._class,
     _key: `snyk-project-finding-${vuln.id}`,
-    _type: SNYK_FINDING_ENTITY_TYPE,
+    _type: Entities.SNYK_FINDING._type,
     category: 'application',
-    score: vuln.cvssScore,
-    cvssScore: vuln.cvssScore,
-    cwe: vuln.identifiers.CWE,
-    cve: vuln.identifiers.CVE,
-    description: vuln.description,
-    displayName: vuln.title,
-    webLink: vuln.url,
+    score: vuln.issueData.cvssScore,
+    cvssScore: vuln.issueData.cvssScore,
+    cwe: vuln.issueData.identifiers?.CWE,
+    cve: vuln.issueData.identifiers?.CVE,
+    description: vuln.issueData.description,
+    displayName: vuln.issueData.title,
+    webLink: vuln.issueData.url,
     id: vuln.id,
-    numericSeverity: vuln.cvssScore,
-    severity: vuln.severity,
-    from: vuln.from,
-    package: vuln.package,
-    version: vuln.version,
-    language: vuln.language,
-    packageManager: vuln.packageManager,
-    isUpgradable: vuln.isUpgradable,
-    isPatchable: vuln.isPatchable,
-    publicationTime: getTime(vuln.publicationTime),
-    disclosureTime: getTime(vuln.disclosureTime),
+    numericSeverity: vuln.issueData.cvssScore,
+    severity: vuln.issueData.severity,
+    pkgName: vuln.pkgName,
+    pkgVersions: vuln.pkgVersions,
+    language: vuln.issueData.language,
+    isUpgradable: vuln.fixInfo?.isUpgradable,
+    isPatchable: vuln.fixInfo?.isPatchable,
+    publicationTime: parseTimePropertyValue(vuln.issueData.publicationTime),
+    disclosureTime: parseTimePropertyValue(vuln.issueData.disclosureTime),
     open: true,
     targets: [],
-    type: vuln.type,
+    issueType: vuln.issueType,
     identifiedInFile: '',
   };
 }
 
-export function createCVEEntity(cve: string, cvssScore: number): Entity {
+export function createCVEEntity(
+  cve: string,
+  cvssScore: number | string,
+): CVEEntity {
   const cveLowerCase = cve.toLowerCase();
   const cveUpperCase = cve.toUpperCase();
   const link = CVE_URL_BASE + cveUpperCase;
   return {
-    _class: 'Vulnerability',
+    _class: Entities.CVE._type,
     _key: cveLowerCase,
-    _type: SNYK_CVE_ENTITY_TYPE,
+    _type: Entities.CVE._class,
     name: cveUpperCase,
     displayName: cveUpperCase,
     cvssScore: cvssScore,
@@ -130,16 +79,16 @@ export function createCVEEntity(cve: string, cvssScore: number): Entity {
   };
 }
 
-export function createCWEEntity(cwe: string): Entity {
+export function createCWEEntity(cwe: string): CWEEntity {
   const cweLowerCase = cwe.toLowerCase();
   const cweUpperCase = cwe.toUpperCase();
   const link = `https://cwe.mitre.org/data/definitions/${
     cwe.split('-')[1]
   }.html`;
   return {
-    _class: 'Weakness',
+    _class: Entities.CWE._class,
     _key: cweLowerCase,
-    _type: SNYK_CWE_ENTITY_TYPE,
+    _type: Entities.CWE._type,
     name: cweUpperCase,
     displayName: cweUpperCase,
     references: [link],
@@ -154,7 +103,7 @@ export function createServiceFindingRelationship(
   return {
     _class: 'IDENTIFIED',
     _key: `${service._key}|identified|${finding._key}`,
-    _type: SNYK_SERVICE_SNYK_FINDING_RELATIONSHIP_TYPE,
+    _type: Relationships.SERVICE_IDENTIFIED_FINDING._type,
     _fromEntityKey: service._key,
     _toEntityKey: finding._key,
     displayName: 'IDENTIFIED',
@@ -163,12 +112,12 @@ export function createServiceFindingRelationship(
 
 export function createFindingVulnerabilityRelationship(
   finding: Entity,
-  cve: Entity,
+  cve: CVEEntity,
 ): Relationship {
   return {
     _key: `${finding._key}|is|${cve._key}`,
     _class: 'IS',
-    _type: SNYK_FINDING_CVE_RELATIONSHIP_TYPE,
+    _type: Relationships.FINDING_IS_CVE._type,
     _mapping: {
       sourceEntityKey: finding._key,
       relationshipDirection: RelationshipDirection.FORWARD,
@@ -181,12 +130,12 @@ export function createFindingVulnerabilityRelationship(
 
 export function createFindingWeaknessRelationship(
   finding: Entity,
-  cwe: Entity,
+  cwe: CWEEntity,
 ): Relationship {
   return {
     _key: `${finding._key}|is|${cwe._key}`,
     _class: 'EXPLOITS',
-    _type: SNYK_FINDING_CWE_RELATIONSHIP_TYPE,
+    _type: Relationships.FINDING_EXPLOITS_CWE._type,
     _mapping: {
       sourceEntityKey: finding._key,
       relationshipDirection: RelationshipDirection.FORWARD,